requirements.txt
pyTelegramBotAPI-3.7.2
CurrencyConverter-0.14.2
requests-2.24.0